Remove map controls and on-map buttons.
- Move "Clear all" map control to left of "Link to last result" as "Clear
  map" link
- Remove map controls bar from above map
- Remove all map controls
  - Find address at center
  - Show bike map
- Remove map "Bike map legend" button from map
- Remove all associated event handlers
- Remove all associated functions
  - byCycle.services.handleMapClick()
  - byCycle.services.identifyIntersectionAtCenter()
  - byCycle.services.identifyIntersection()
  - byCycle.services.identifyStreet()
  - byCycle.services.toggleBikeTileOverlay()
  - byCycle.ui.handleMapClick()
- Remove associated map_controls() Mako <%def>
- Remove associated style definitions

Address lookup functionality map be reintroduced in the future, but, if
so, it will be accessed via a map context menu.

The bike map layer and legend may be reintroduced if I can figure out an
easy way to display the layer (i.e., I don't want to set up or maintain a
map server).