In service controllers, fix how action is used for rendering
When appropriate, set `self.action`; Restler uses self.action to select
which template to render in `Controller._render_template()`. This is done
when multiple results are found for a query or when an error is
encountered.

This also fixes a bug where the action could be wrong when rendering the
fragment for JSON responses.