Deploy static files to separate app dir on WF
This makes it really easy to set up static file serving. Basically, copy
the static dir from the project to the static file app dir on WF and
that's it. Other than clicking a few buttons in the WF control panel to
do the initial setup, no other server config is needed.