Create scripts in production
In case we need to reintegrate or rebuild matrix.

There should be a better way of dealing with this situation. If the data
model changes (or if regional data is updated), we should be able to
reintegrate and rebuild the matrix "offline" and then just dump the data
into production rather than taking the DB offline.