Get version for static_url from settings
Trying to retrieve the hg revision at runtime is hacky. Also, eventually,
we should be using proper version numbers instead of hg revisions, and
this change is forward compatible with that.

For now, this is going to require running the following to set the
version in production:

    buildout -c deploy.cfg wsgi_application:version=${VERSION}

When we start using proper version numbers, we can retrieve the version
automatically at runtime using pkg_resources.