Speed up JS building by using rJSMin instead of uglifyjs
r.js is now used only to assemble a single JS file. rjsmin is then run
on that file.

This is *much* faster than before where r.js ran uglifyjs for us. That
required adjusting Node's stack size because it was so memory intensive
and was painfully slow (which discouraged rapid/continuous) deployments.