Revert back to Python 3.5
Python 3.5 is installed on the production server and the dev environment
should match that to avoid accidentally using backward-incompatible
features. The plan is to upgrade to Python 3.6 after Ubuntu 18.04 LTS
comes out.

The code changes involve just not using `f'XYZ'` strings. This also
reverts the Python version classifier and removes `python.version` from
the RunCommands config (bycycle.core now sets this to 3.5 for all
environments).