Revamp deployment configuration
The previous AWS configuration was too complex (and expensive), so this
scales back to a single EC2 instance (via Lightsail) running Nginx and
uWSGI that talks to RDS. It uses Ubuntu's standard Nginx and uWSGI
setups (sites-available and apps-available). One thing that's nice about
this is that we don't need to mess with init scripts for uWSGI. The
info.bycycle.org config was also removed as that now lives in a separate
repo.