Use Cornice to configure web services
This replaces use of Pyramid's lower level route and view registration
with Cornice resource registration.

[I had initially started working on an updated version of
wylee/pyramid_restler, but Cornice is close enough to what I was trying
to accomplish with that and is also more mature.]

This also splits the query facade out of the base service view/resource
into a separate view/resource.